+++ R/dilplots-plotly.R
@@ -377,56 +377,67 @@ plot_curve_plotly <- function(
   return(p)
 }
 
-
 #' @title Add A `plotly` Panel Column
-#' @description Create a column which contains a list of `plotly
-#' suited for a `trelliscopejs` visualisation
-#' @param dilution_table Output given from
-#' the function [create_curve_table()]
+#' @description Create a column which contains a list of `plotly`
+#' plots suited for a `trelliscopejs` visualisation.
+#' @param curve_table Output given from
+#' the function [create_curve_table()].
 #' It is in long table format with columns indicating at least the
 #' lipid/transition name, the concentration and signal. Other columns may be
-#' present if it is used to group the dilution curve together
-#' @param dilution_summary The summary table generated
+#' present if it is used to group the curve together.
+#' @param curve_summary The summary table generated
 #' by function [summarise_curve_table()] and/or
 #' [evaluate_linearity()]
-#' but it can also be any generic data frame or tibble
+#' but it can also be any generic data frame or tibble.
 #' Default: NULL
+#' @param dilution_table `r lifecycle::badge("deprecated")`
+#' `dilution_table` was renamed to
+#' `curve_table`.
+#' @param dilution_summary `r lifecycle::badge("deprecated")`
+#' `dilution_summary` was renamed to
+#' `curve_summary`.
 #' @param grouping_variable A character vector of
-#' column names in `dilution_table`to indicate how each dilution curve
+#' column names in `curve_table`to indicate how each curve
 #' should be grouped by. It is also going to be used as a conditional
 #' cognostics in the `trelliscopejs` report.
 #' Default: c("Transition_Name", "Dilution_Batch_Name")
-#' @param sample_name_var Column name in `dilution_table`
-#' to indicate the sample name. To be used in the dilution plot
+#' @param sample_name_var Column name in `curve_table`
+#' to indicate the sample name. To be used in the curve plot.
 #' Default: 'Sample_Name'
-#' @param dil_batch_var Column name in `dilution_table`
-#' to indicate the group name of each dilution batch,
-#' used to colour the points in the dilution plot
+#' @param curv_batch_var Column name in `curve_table`
+#' to indicate the group name of each curve batch,
+#' used to colour the points in the curve plot.
 #' Default: 'Dilution_Batch_Name'
-#' @param dil_batch_col A vector of colours to be used for the dilution
-#' batch group named given in `dil_batch_var`,
+#' @param curv_batch_col A vector of colours to be used for the curve
+#' batch group named given in `curv_batch_var`.
 #' Default: c("#377eb8", "#4daf4a", "#9C27B0", "#BCAAA4", "#FF8A65", "#EFBBCF")
-#' @param conc_var Column name in `dilution_table` to indicate concentration
+#' @param dil_batch_var `r lifecycle::badge("deprecated")`
+#' `dil_batch_var` was renamed to
+#' `curv_batch_var`.
+#' @param dil_batch_col `r lifecycle::badge("deprecated")`
+#' `dil_batch_col` was renamed to
+#' `curv_batch_col`.
+#' @param conc_var Column name in `curve_table` to indicate concentration.
 #' Default: 'Dilution_Percent'
-#' @param conc_var_units Unit of measure for `conc_var`, Default: '%'
+#' @param conc_var_units Unit of measure for `conc_var`. Default: '%'
 #' @param conc_var_interval Distance between two tick labels
-#' in the dilution plot,
+#' in the curve plot.
 #' Default: 50
-#' @param signal_var Column name in `dilution_table` to indicate signal
+#' @param signal_var Column name in `curve_table` to indicate signal.
 #' Default: 'Area'
 #' @param have_plot_title Indicate if you want to have a plot title in
 #' the `plotly` plot.
 #' Default: FALSE
 #' @param plot_first_half_lin_reg Decide if we plot an extra regression line
-#' that best fits the first half of `conc_var` dilution points.
+#' that best fits the first half of `conc_var` curve points.
 #' Default: FALSE
 #' @param plot_last_half_lin_reg Decide if we plot an extra regression line
-#' that best fits the last half of `conc_var` dilution points.
+#' that best fits the last half of `conc_var` curve points.
 #' Default: FALSE
 #' @return A table that is suited for a `trelliscopejs` visualisation with
 #' `grouping variable` columns converted to conditional cognostics,
-#' other columns in `dilution_summary` converted to cognostics and
-#' a new column `panel` created containing a `plotly` dilution plot in each row.
+#' other columns in `curve_summary` converted to cognostics and
+#' a new column `panel` created containing a `plotly` curve plot in each row.
 #' This column is used to create the plot figure in the
 #' `trelliscopejs` visualisation.
 #' @examples
@@ -529,7 +540,7 @@ plot_curve_plotly <- function(
 #' # Create a trellis table
 #' trellis_table <- add_plotly_panel(
 #'   curve_table,
-#'   dilution_summary = curve_summary
+#'   curve_summary = curve_summary
 #' )
 #'
 #' plotly_list <- trellis_table$panel
@@ -542,46 +553,75 @@ plot_curve_plotly <- function(
 #'
 #' @rdname add_plotly_panel
 #' @export
-add_plotly_panel <- function(dilution_table, dilution_summary = NULL,
-                             grouping_variable = c(
-                               "Transition_Name",
-                               "Dilution_Batch_Name"
-                             ),
-                             sample_name_var = "Sample_Name",
-                             dil_batch_var = "Dilution_Batch_Name",
-                             dil_batch_col = c(
-                               "#377eb8",
-                               "#4daf4a",
-                               "#9C27B0",
-                               "#BCAAA4",
-                               "#FF8A65",
-                               "#EFBBCF"
-                             ),
-                             conc_var = "Dilution_Percent",
-                             conc_var_units = "%",
-                             conc_var_interval = 50,
-                             signal_var = "Area",
-                             have_plot_title = FALSE,
-                             plot_first_half_lin_reg = FALSE,
-                             plot_last_half_lin_reg = FALSE) {
+add_plotly_panel <- function(
+    curve_table,
+    curve_summary = NULL,
+    dilution_table = lifecycle::deprecated(),
+    dilution_summary = lifecycle::deprecated(),
+    grouping_variable = c("Transition_Name",
+                          "Dilution_Batch_Name"),
+    sample_name_var = "Sample_Name",
+    curv_batch_var = "Dilution_Batch_Name",
+    curv_batch_col = c("#377eb8", "#4daf4a",
+                       "#9C27B0", "#BCAAA4",
+                       "#FF8A65", "#EFBBCF"),
+    dil_batch_var = lifecycle::deprecated(),
+    dil_batch_col = lifecycle::deprecated(),
+    conc_var = "Dilution_Percent",
+    conc_var_units = "%",
+    conc_var_interval = 50,
+    signal_var = "Area",
+    have_plot_title = FALSE,
+    plot_first_half_lin_reg = FALSE,
+    plot_last_half_lin_reg = FALSE) {
 
+  if (lifecycle::is_present(dilution_table)) {
+    lifecycle::deprecate_warn(
+      when = "0.0.6.9000",
+      what = "add_plotly_panel(dilution_table)",
+      with = "add_plotly_panel(curve_table)")
+    curve_table <- dilution_table
+  }
 
+  if (lifecycle::is_present(dilution_summary)) {
+    lifecycle::deprecate_warn(
+      when = "0.0.6.9000",
+      what = "add_plotly_panel(dilution_summary)",
+      with = "add_plotly_panel(curve_summary)")
+    curve_summary <- dilution_summary
+  }
 
-  # Check if dilution_table is valid with the relevant columns
+  if (lifecycle::is_present(dil_batch_var)) {
+    lifecycle::deprecate_warn(
+      when = "0.0.6.9000",
+      what = "add_plotly_panel(dil_batch_var)",
+      with = "add_plotly_panel(curv_batch_var)")
+    curv_batch_var <- dil_batch_var
+  }
+
+  if (lifecycle::is_present(dil_batch_col)) {
+    lifecycle::deprecate_warn(
+      when = "0.0.6.9000",
+      what = "add_plotly_panel(dil_batch_col)",
+      with = "add_plotly_panel(curv_batch_col)")
+    curv_batch_col <- dil_batch_col
+  }
+
+  # Check if curve_table is valid with the relevant columns
   validate_curve_table(
-    curve_table = dilution_table,
+    curve_table = curve_table,
     needed_column = c(
       grouping_variable,
       sample_name_var,
-      dil_batch_var,
+      curv_batch_var,
       conc_var,
       signal_var
     )
   )
 
-  # Try to create dilution summary if you do not have one.
-  if (is.null(dilution_summary)) {
-    dilution_summary <- dilution_table %>%
+  # Try to create curve summary if you do not have one.
+  if (is.null(curve_summary)) {
+    curve_summary <- curve_table %>%
       summarise_curve_table(
         grouping_variable = grouping_variable,
         conc_var = conc_var,
@@ -590,26 +630,26 @@ add_plotly_panel <- function(dilution_table, dilution_summary = NULL,
       evaluate_linearity(grouping_variable = grouping_variable)
   }
 
-  # Check if things in needed_column are in dilution_summary
-  assertable::assert_colnames(dilution_summary, grouping_variable,
+  # Check if things in needed_column are in curve_summary
+  assertable::assert_colnames(curve_summary, grouping_variable,
     only_colnames = FALSE, quiet = TRUE
   )
 
-  # Get the dilution batch name from dilution_table
-  dilution_batch_name <- dilution_table %>%
-    dplyr::pull(.data[[dil_batch_var]]) %>%
+  # Get the curve batch name from curve_table
+  curve_batch_name <- curve_table %>%
+    dplyr::pull(.data[[curv_batch_var]]) %>%
     unique() %>%
     as.character()
 
-  # Create palette for each dilution batch for plotting
-  pal <- dil_batch_col %>%
-    create_char_seq(output_length = length(dilution_batch_name)) %>%
-    stats::setNames(dilution_batch_name)
+  # Create palette for each curve batch for plotting
+  pal <- curv_batch_col %>%
+    create_char_seq(output_length = length(curve_batch_name)) %>%
+    stats::setNames(curve_batch_name)
 
 
   # Create a title name for each group
   if (isTRUE(have_plot_title)) {
-    dilution_table <- dilution_table %>%
+    curve_table <- curve_table %>%
       dplyr::rowwise() %>%
       dplyr::mutate(title = paste0(
         dplyr::across(dplyr::all_of(grouping_variable)),
@@ -617,14 +657,14 @@ add_plotly_panel <- function(dilution_table, dilution_summary = NULL,
       )) %>%
       dplyr::ungroup()
   } else {
-    dilution_table <- dilution_table %>%
+    curve_table <- curve_table %>%
       dplyr::mutate(title = "")
   }
 
-  # Add dil_batch_var in the nested data
-  # Will not work if dil_batch_var is also a grouping_variable
+  # Add curv_batch_var in the nested data
+  # Will not work if curv_batch_var is also a grouping_variable
 
-  dilution_table <- dilution_table %>%
+  curve_table <- curve_table %>%
     dplyr::group_by_at(
       c(grouping_variable, "title")
     ) %>%
@@ -634,22 +674,22 @@ add_plotly_panel <- function(dilution_table, dilution_summary = NULL,
   # If this is the case, we need to make a copy
   # of the variable inside the nested data
 
-  if (dil_batch_var %in% grouping_variable) {
-    dilution_table <- dilution_table %>%
+  if (curv_batch_var %in% grouping_variable) {
+    curve_table <- curve_table %>%
       dplyr::mutate(data = purrr::map2(
         .x = .data$data,
-        .y = .data[[dil_batch_var]],
-        .f = function(df, dilution_batch_name) {
+        .y = .data[[curv_batch_var]],
+        .f = function(df, curve_batch_name) {
           df <- df %>%
-            dplyr::mutate(!!dil_batch_var := dilution_batch_name)
+            dplyr::mutate(!!curv_batch_var := curve_batch_name)
           return(df)
         }
       ))
   }
 
-  # Group/Nest the dilution data for each group
-  # and do a dilution plot for each of them
-  dilution_plots <- dilution_table %>%
+  # Group/Nest the curve data for each group
+  # and do a curve plot for each of them
+  curve_plots <- curve_table %>%
     dplyr::ungroup() %>%
     dplyr::mutate(panel = trelliscopejs::map2_plot(
       .x = .data$data,
@@ -657,7 +697,7 @@ add_plotly_panel <- function(dilution_table, dilution_summary = NULL,
       .f = plot_curve_plotly,
       pal = pal,
       sample_name_var = sample_name_var,
-      curv_batch_var = dil_batch_var,
+      curv_batch_var = curv_batch_var,
       conc_var = conc_var,
       conc_var_units = conc_var_units,
       conc_var_interval = conc_var_interval,
@@ -667,12 +707,12 @@ add_plotly_panel <- function(dilution_table, dilution_summary = NULL,
     ))
 
 
-  # Left Join plots with grouping variable and dilution_summary
-  trellis_table <- dilution_plots %>%
+  # Left Join plots with grouping variable and curve_summary
+  trellis_table <- curve_plots %>%
     dplyr::select(dplyr::all_of(c(grouping_variable))) %>%
-    dplyr::bind_cols(dilution_plots %>%
+    dplyr::bind_cols(curve_plots %>%
       dplyr::select(dplyr::any_of("panel"))) %>%
-    dplyr::left_join(dilution_summary, by = grouping_variable) %>%
+    dplyr::left_join(curve_summary, by = grouping_variable) %>%
     dplyr::relocate(
       dplyr::any_of("panel"),
       .after = dplyr::last_col()